move this to its own file and add documentation
